+++ packages/graphql/src/GraphQL-FAQ.md
