+++ packages/graphql/src/constants/projectConstants.ts
