+++ packages/graphql/src/entities/App.ts
