+++ packages/graphql/src/entities/WizardFrontendFramework.ts
