+++ packages/graphql/src/entities/WizardNpmPackage.ts
